add ssr to events
